backend/requirements.txt
altair==5.5.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
boto3==1.40.45
botocore==1.40.45
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-sendgrid-v5==1.3.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.6.0
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mysqlclient==2.2.7
narwhals==2.4.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
seaborn==0.13.2
sendgrid==6.12.5
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3       
whitenoise     
